Make default typeface return empty
This updates all of our tests and gms to behave like they did before.
Many of our gms (and some of our tests) use the portable typeface
and font, so I made that explicit instead of relying on the
default font (with a null typeface) to be resolved in that way.

After talking with bungeman@, I added ToolUtils::DefaultTypeface()
and similar functions because in our tests, we probably want to
use the actual typefaces loaded via the compiled-in font manager,
and this will give us a mechanism to do so after SkFontMgr::RefDefault()
goes away.

This should not affect clients (yet), as I'm working with them to
set the define for them and resolve the breakages.

Change-Id: I9ebfce2d8c17e34cb837b4bbd7e0345c9e5f513d
Bug: b/305780908
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/768618
Owners-Override: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>