[canvaskit] Get rid of dependency on default fontmgr
CanvasKit does not really need a Font Manager because there are no
system fonts to manage. The previous exception to this was the
embedded font that CanvasKit ships with. Given that we aim to
remove the default font manager, now seemed like a good time to
bypass that while still preserving the existing behavior.

Removing FontMgr entirely from the API seems a bit heavy handed -
clients may want to pass in a bunch of fonts and then look them
up, which seems like a perfectly valid use to keep it around in
that capacity.

Change-Id: I256869031cbe0ede29d451287f1663cda2d24f84
Bug: b/305780908
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/770357
Reviewed-by: Ben Wagner <bungeman@google.com>