Add drawing downsampler
- tiny code-size, since it reuses the drawing code
- filtering may not be as fine-tuned as the HQ downsampler

This downsampler is not enabled by default (since it slightly changes gms) but is checked in so it can be experimented with. Might consider using build or runtime flags to switch between this and HQ while we figure out which clients want which one.

Bug: b/307987911
Change-Id: I40e370a9b338f0668366a3feb2ebb14c1849c209
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/771699
Commit-Queue: Mike Reed <mike@reedtribe.org>
Auto-Submit: Mike Reed <mike@reedtribe.org>
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>