Graphite: calculate UBO/SSBO's binding size in DrawPass::writeUniforms.
The binding size is important for backend such as Dawn to create
BindGroup.

Previously the binding size is calculated in Dawn backend by extracting
the uniforms' total bytes from the generated shader source via
EmitPaintParamsUniforms or EmitRenderStepUniforms. However for SSBO,
this might not be possible since the SSBO could be unsized in shader
source.

The more correct way of calculating binding size of UBO/SSBO could be
doing it in DrawPass::writeUniforms().

Bug: b/260368758
Change-Id: I4f1b1d5511970a24b05ba4d4925e704a76559977
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/772837
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Quyen Le <lehoangquyen@chromium.org>