Graphite: Store wgpu::TextureAspect and create plane WGPUTextureView
Add a new BackendTexture constructor that takes in WGPUTexture and
DawnTextureInfo. The DawnTextureInfo also takes in a
wgpu::TextureAspect which is also passed to DawnTextureSpec. This
aspect and WGPUTexture can then be used to create a plane
WGPUTextureView when creating a wrapped DawnTexture. This allows us
to later pass the TextureAspect for copy operations as well through
the DawnCommandBuffer.

Note: A follow-up change will remove the deprecated BackendTexture
ctor taking in WGPUTextureView once Chrome side change starts using the
new ctor. This can also clean up storing the fDawnTextureView and
DawnTexture methods.

Bug: b/284398622
Change-Id: Ie622b3f0a942182120ac5203dedeeaffd6bae0e4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/773116
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>