Fix fuzzer-discovered assertion with duplicate symbols.
We don't have a good mechanism to detect that `SymbolTable::add`
has rejected a symbol because it is a duplicate; code was always
written assuming that an `add` will be successful, and it didn't
matter what happened if that assumption was wrong, because we had
reported an error and compilation was going to stop soon.
However, because followup code will assume that the symbol
is there and has the proper type, the fuzzer was able to
construct a contradiction that led to an assert.

Because followup code is going to assume that adding succeeds,
adding a duplicate symbol now replaces the existing symbol in
the symbol table (while reporting an error, as always). This
solves the fuzzer test case and doesn't change any results
from the rest of our test suite. There's a small chance that
the fuzzer will use this opportunity to find a new contradiction
to exploit; if that happens we'll probably need to update
SymbolTable::add to return success/failure.

Bug: oss-fuzz:63793
Change-Id: I71da312e280e3271f743add919b818f290917e16
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/773837
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>