Find Segments that round to the same point
Part of Bentley-Ottmann is finding all the segments that
pass through a given point on the sweep line. The
problem is that we need to find the segments that round
to a given point on the sweep line because they might
not pass directly through the integer coordinate.

The rounded segments are given by:
x = ⌊s(y) + ½⌋

We can break this into two comparisons:
(x - ½) ≤ s(y) < (x + ½)

Where (x, y) are the given point, and s(y) is the
function for the segment that returns x given y.

Implement the two halves of the inequality above as
two separate functions.
   rounded_point_less_than_segment_in_x_lower
and
   rounded_point_less_than_segment_in_x_upper

An example of how to use them to find the rounded
segments on the sweep line is given in SweepLineTest.cpp

Bug: b/297220808
Change-Id: I8d92cdfc8f22987f5e488f240fca136edfeb71de
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/774136
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Herb Derby <herb@google.com>