[graphite] Skip calling getMostRecentDraw() for draws that don't need it for sorting
Removes `fDrawsOverlap` and its tracking since that was the only use
of getMostRecentDraw() when a draw was opaque and used MSAA, but
there's no logic yet to take advantage of `fDrawsOverlap`. Down the
line, we should probably only be tracking draw overlapping when on
an offscreen device for a layer whose restore paint is known to be
compatible with elision anyways.

Bug: b/238758897
Change-Id: I61fa01c7786eb4d6019eff8af8e0aae2d4ec8610
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/774576
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>