[ganesh] Fix dashed circle calculations near 2*pi
Before this CL, we could drop "next" dashes within an interval if the
interval ended at 2*pi, due to floating point error in calculations.
This CL introduces an epsilon into a comparison when deciding to drop
the next dash. It uses 0.01, interpreting an existing use of this
constant as an epsilon when deciding to drop the previous dash.

Bug: chromium:1495670
Change-Id: I127555d20e099915e484daef5be9bb8b711c0b4b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/774696
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>