[graphite] Add upper bound to grid size for Grid/HybridBoundsManagers
The GridBoundsManager (and transitively the HybridBoundsManager when
there are lots of draws) is sensitive to the grid size used to cover
a surface. The MakeRes factory can create an arbitrarily large grid
to preserve the target pixel cell size, if the surface it's covering
is also really large.

The performance characteristics change pretty radically on motionmark
when running on a high-dpi surface vs. a regular resolution surface
because of this. Add an upper bound to the grid size to avoid
this cliff.

This significantly reduces the CPU overhead of the bounds manager
tracking in motionmark scenes (4x on a high-dpi config benchmark
emulation), but slightly increases the number of draw batches on
the scenes (arcs goes from 660 to 710). nanobench was still
reporting a 5-15% improvement for net time on motionmark arcs
and paths using the `grmtl` config.

Change-Id: I1290192f8ee768b38af1353c91c88c72445602f9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/774356
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>