GrBlurUtils: Subset on low-res pixel boundary, not float coords
The low-res blurred image was rendered into the rounded-out coords,
and any decal-tiling should occur beyond their boundaries. By using
the original scaled coords as the subset boundary, we'd introduce
transparency at the edge when re-expanding.

Previously this was never visible because of quirks in how
SkBlurImageFilter requested its output (always padded with the blur
radius, even if it were to then be cropped).

While the changes to the subsetting logic accounted for 90% of the
visible seams when testing in chrome (and all the visible seams when
running in viewer), it was also necessary for the downscaled blur
to write one extra pixel of padding that could be sampled from when
doing the re-expand.

Bug: chromium:1500021
Change-Id: I18c147270db7b42568f4e7034b13c2dc4ab19e4a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/775517
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>