Force coloremoji GM to have a small multitexture atlas.
There are a couple of reasons for this. First, when this GM is run with
color transformations this verifies that the bug below continues to be
fixed. Second, we're currently only doing this kind of stress test on
the mask atlas, so it's reasonable to verify color text as well.

Bug: b/309422936
Change-Id: I9995ad6f454bcecb21ea4f56d490b2c0edbabdb9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/774569
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>