Graphite: Clean up BackendTexture fDawnTextureView and related methods
This change cleans up deprecated BackendTexture constructor and
dawnTextureView related code that is no longer needed/used.

Also, rename some of the passed in params as planeDimensions and
viewFormat to make it clearer that they are always for a plane/view for
multiplanar textures and don't correspond to the size/format of the
texture.

Bug: b/284398622
Change-Id: If9a37feefa28af3d818f178855b72afd0aa3195f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/773943
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>