Add SkFontMgr option to DataURI resources and clean up predecoding
I've been bitten before by having a boolean parameter being incorrectly
set by a pointer, so this changes the "predecode" boolean to
be a class enum to make the addition of the SkFontMgr less
error-prone.

Looking at current clients who use DataURIResourceProviderProxy,
they might only be relying on the base64 decoding of images,
not typeface data. Therefore, we might be able to land the SkFontMgr
part of this change w/o staging.

Also clients who use MultiFrameImageAsset, FileResourceProvider,
and DataURIResourceProviderProxy are not setting the predecode
boolean, so that also can probably be landed w/o staging.

The increase in code size is expected, given the fact that previously
skottie_tool had the empty SkFontMgr instead of an actual SkFontMgr
which could load fonts (Freetype was previously included to deserialize
typefaces, I believe)

Change-Id: I99c782c71a4b24479f3ec9d3f226a6f37e2101b7
Bug: b/305780908
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/773576
Reviewed-by: Florin Malita <fmalita@google.com>