RasterPathAtlas: fix key builder index typo
Currently, fracBits and styleBits are stored in builder[6], which will
immediately be overwritten by shape.writeKey(&builder[6], ...) on the
next line. builder[5] seems to be unused.

It looks like the intention was to store fracBits and styleBits in
builder[5], so let's do that.

Change-Id: I7d31513b89dfe94a7859957b2b4d8425efc70aac
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/777336
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>