Update GrBufferAllocPool pushBack to not loop over or destroy blocks.
This hopefully fixes the linked bug/crash. The idea is that the pushBack
is causing a block to be destroyed which makes a previous block the
new top of the stack. However, the fBufferPtr doesn't get updated to
point back to this old block and thus the crash. We don't actually
think that pushBack should ever need to return bytes outside of the
current block so we removed the loop. Additionally if the returned
bytes was exactly the size of the block we would destroy that block
instead of reusing the whole thing.

With this change we shouldn't end up in a situation where we have
a null fBufferPtr but have a block in the stack.

Bug: chromium:1500788
Change-Id: I5d08f7b78fe59b115c5bf39811acc8ce87133a5c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/777321
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>