Avoid recursion into the fields of a StructType.
Previously, if we needed to know the slot count of a struct, or
whether it contained arrays, we would compute it recursively.
This could be expensive for nested structs. The fuzzer discovered
it could make extremely deeply-nested trees of structs and cause
a timeout. We now precompute this information for a struct as
soon as it is constructed; after this change, the fuzz test
fails immediately.

Bug: oss-fuzz:64113
Change-Id: I32a2543b1ee1ce94627c8f35f36f5cd1893c50d9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/778141
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>