Graphite: Relax wgpu::TextureAspect validation
Relax TextureAspect validation that is not known for creating promise
images and replay recordings. For such cases we can just pass
TextureAspect::All and do not have to be Plane0/1. This relaxes the
validation so during BackendTexture creation the correct aspect is
passed in to create the WGPUTextureView but for promise images Aspect
all is allowed.

Bug: chromium:1500366
Change-Id: I7e9316c7b85c2b38efbdaca237bd2828575e433a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/777037
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>