Add buffet rule for png encoding and fontconfig SkFontMgr
Clients need a way to get @freetype_config (a small set of files
to configure how freetype is compiled), so I added
bazel/download_config_files.bzl to copy them from the Skia repository.

We will eventually want something more robust/hermetic, but for now
this just downloads them from GitHub because that was easiest.

When using a toolchain that is not the Skia hermetic one, clients
will have to have the fontconfig headers and something to link against
(these are in the docker container we test the compilation with [1]).

[1] https://github.com/google/skia/blob/3e1e653634557386e53eaecbca97f6489d25c990/infra/gcc/Debian11/Dockerfile#L6

Change-Id: I5b16f98a487091ea95916946b9094eb1a22b71ae
Bug: b/305780908
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/778139
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>