[skif] Adjust blur engine tile mode support API
This restricts it so that findAlgorithm() does not see the tile mode
that will be applied, and an algorithm's support is restricted to
either "all tilemodes" or "just-decal".

This makes image prep and resizing logic easier once FilterResult
implements tile-aware rescaling because it can either apply the tile
mode or not, but sometimes wants to switch to clamp-tiling if it can.
With this change there's no need for blur() to requery an algorithm,
which has its own complications to not get into.

Bug: b/294102201
Change-Id: I9d852c9353fde5dd349842d8888dcc9b638de3f6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/780216
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>