SkSVGDevice: don't skip fill-opacity when fill is black
Commit 05bab9a29d621cb155c4fba80d7e8bb50927bdd7 introduced a regression
that caused the fill-opacity attribute to not be omitted when the
svg paths have a black (default) fill.

The two attributes are orthogonal and the fact that the fill is black
should not cause the fill-opacity attribute to not be emitted.
This is the case when we draw a black shape that is semi (or fully)
transparent.

Bug: b/309068101
Change-Id: Ic0bef56404d5369f486552da03b82348037caf5f
Signed-off-by: Joey Bevilacqua <bevilacquajoey@gmail.com>
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/780277
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>