Remove other uses of SkFontMgr::RefDefault
Clients of Skottie should be setting their own SkFontMgrs and not
be relying on the default, so we don't need to lazily load it.

SkParagraph already has a fallback fontmgr, so let's use that
instead of the default.

SkShaper_harfbuzz appears to only use the fontmgr for fallback,
which I documented some and removed the behavior where a null
fallback SkFontMgr would default to SkFontMgr::RefDefault()
(it now means the empty FontMgr or "no" fallback).

After this CL lands, we still need to clean up clients.

Bug: b/305780908
Change-Id: If7e9dd148e561303a424bca823e4d7c1bc5b9183
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/779396
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>