Remove MutableTextureStateRef and sketch out Vulkan-specific functions
The goal is to remove the Vulkan #ifdefs from MutableTextureState
and while doing that, we realized the extra layer of indirection was
not necessary. This makes MutableTextureState inherit from SkRefCnt
so it can go in a sk_sp directly.

As long as I was cleaning up function calls (they were named slightly
differently on State and StateRef), I added the Vulkan-only file and
a skgpu::MutableTextureStates namespace for them to live in. The
implementation of this will change in a follow-up CL when we remove
the type union from MutableTextureState.

Bug: b/293490566
Change-Id: Ic7178a9bd7a0be19674113a4becc308ac8774c23
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/781498
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>