Remove type union from MutableTextureState
The original VulkanMutableTextureState had two atomic fields,
but we cannot copy an atomic (e.g. in VulkanMutableTextureState::copyTo).

The atomic on the old thing wasn't really protecting us from
much because there were two individual fields. We might have
added them to try to guard the clients from mutating them,
but that would be a recipe for disaster anyway.

In Graphite, we need to be sure not to try to read this
from a Recorder thread because we will be mutating it on
the context thread. An atomic wouldn't really help us there
either, so we have omitted the atomic in this impl.

Change-Id: I390105b42d246c73bee337840ef6b6895d5b0372
Bug: b/293490566
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/782417
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>