Increase SkQP runtime-hint parameter to 20m
SkQP is still exceeding 15m on CF in a number of TEST_MAPPING runs
(~12% over the past 5 days). This is a shot in the dark, but I'm hoping
this might allow better sharding decisions to be made for this module.
I haven't found much documentation for it yet, but it doesn't look
dangerous IMO. :)

Test: analyze history with ATH after submitting (testmappingsub:external/skia)
Bug: b/313936517
Change-Id: If65534ed56840e8967208c6450810644b759d4e0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/783999
Commit-Queue: Nolan Scobie <nscobie@google.com>
Reviewed-by: Leon Scroggins <scroggo@google.com>