[skdiff] Move to component build
Since tool_utils is not ready to move to a component build, this
requires some slight modifications to not depend on it. This change also
removes the ability to read in any format other than .png.

This also modifies SkCodec factories to take an explicit list of
decoders. The old factories are retained for backwards compatibility and
continue to use the globals.

To make it easier for users to create the list of decoders,
SkCodecs::Decoders::id is changed from std::string to std::string_view
and the Sk*Decoder::Decoder factories are made constexpr.

This also allows //:tools access to :base and :core_priv.

Change-Id: Id9629dd68d29757a89bf01cce555921b1f577753
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/784577
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>