PredrawFlags and automaskfilter skipping (for text)
In the in-progress automaskfilter system, text is not meant to use a
layer for the mask filter initially. Small text applies the mask
filter directly and saves that in the strike cache. Only when the
text functions in SkDevice route back to SkCanvas for non-masked
rendering should we apply the auto mask filtering.

This restores strike-cached mask filters to small text in graphite,
which is the only user of automask filters at the moment. Initially,
all text with mask filtering would trigger an auto layer.

As part of this, the SkCanvas internal APIs were cleaned up.
ShaderOverrides and CheckForOverdraw enums are combined into a new
PredrawFlags enum bitmask. This new bit mask also has an option for
skipping mask filtering in auto layers. The aboutToDraw(),
predrawNotify(), and other functions now take one enum bitmask instead
of multiple parameters.

This also removes the redundant SkCanvas* parameter from aboutToDraw()
which was always just "this" anyways.

Bug: b/308509317
Bug: b/40042366
Change-Id: Ia796e558487a3120d2ab9d843880da45bd498ac7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/784597
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>