[pdf] Emit color space of images
This also makes Skia PDF directly depend on the Skia JPEG encoders and
decoders instead of making them build time optional. All known users of
PDF already depend on these so it should not make a difference in
practice. If there is a need to allow other (or no) encoders and
decoders that could be considered as a follow up.

Bug: chromium:999986,chromium:1465627
Change-Id: I4992b0d9267a30154cd176528fcb830a39a12494
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/784553
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>