Implement onMatchFamilyStyle for SkParagraph.
CanvasKit uses skparagraph's TypefaceFontProvider which is fine and
dandy but it exposes a very limited interface to FontMgr, with
matchFamilyStyle as the only option for finding typefaces and fonts.

Without implementing this method, it is not possible to use custom fonts
on wasm as calling matchFamilyStyle with current onMatchFamilyStyle is a
noop and returns null.