[graphite] Flush all tracked devices when invalidating PathAtlas.
In certain rare cases we can end up with a child Device flushing its
pending uploads and draws and overwriting a parent's Device atlas data
before the draws associated with that data are flushed. By flushing
all the Devices associated with a Recorder -- similar to how we manage
the text atlas -- we ensure that won't happen.

Bug: chromium:1500498
Change-Id: I83e21923b288eac434d2aab5577df1a636a97b51
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/786568
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>