[skif] Add SkSpecialImages::AsBitmap, remove getROPixels()
The public getROPixels() on SkSpecialImage was never supposed to be
called by non-raster special images, and other than AsView() we could
just switch to a static function to access the SkBitmap.

AsView() no longer uploads raster-backed special images to a texture.
AFIACT this was only used in a test, which has been updated to not
expect that behavior anymore. For real use cases, everything should be
in a consistent gpu/raster world (or mixed backings come from the
regular SkImage).

Bug: b/299474380
Change-Id: I4e0784192e3d17dcb8b89d13983d410dca1e3844
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/787916
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>