Move SKSurfaceProps usage to default constructor
I am planning on adding a new constructor to SKSurfaceProps in a follow
up CL that will add text contrast and gamma parameters. The end
result of this will be two main constructors for SKSurfaceProps:

1) Default/empty constructor with all default values
2) A populated constructor with hardware/system values (pixel geometry,
contrast, gamma, flags, etc.)

There is an existing pattern of mixing these concepts by calling
into a constructor that allows for non-default values, but manually
passing in the defaults, which this CL removes. There should be no
behavioral changes as a result of this CL.

This CL is the corresponding Chromium change:
https://chromium-review.googlesource.com/c/chromium/src/+/5118781

Once that lands, the mixed constructor can be removed entirely.

Bug: chromium:1160653
Change-Id: Ia109b2db49540c6d351d80649bcfb66f5b23b7af
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/789118
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Brian Osman <brianosman@google.com>