[skif] Refine SkSpecialImage::asShader() API for fast constraints
I realized as I was working on the follow-up use of the fast
asShader() that we could decide to use fast sampling on any SkTileMode
if the SkSpecialImage's subset aligned with the backing dimensions and
the tiling was trivially HW.

Additionally, we could leverage HW tiling if just the visible edges of
the SkSpecialImage aligned with the backing dimension. This means we
have to convey the SkTileMode regardless of fast vs. strict, so I just
removed the asShaderFast() and added a parameter to asShader().

Bug: b/315351386
Change-Id: I6e98872ff19f23880426840a0010f57327e31383
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/790078
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>