SkWuffsCodec: Skip copying stream if PreferStillImage
SkWuffsCodec (really the code that creates it) defensively copies the
stream if it's not seekable, so that the codec can seek and therefore
cycle through the animation.

This copy is unnecessary, though, if the client only wants the first
image. Use the SelectionPolicy to determine whether to perform the
copy. Keep track of whether the stream is seekable in SkWuffsCodec.
If not (i.e. the client picked kPreferStillImage so we did not copy
it), report 1 for getFrameCount to avoid seeking through the stream.
This also matches the fact that they are requesting a still image.
Make getFrameInfo fail, too, since that is only useful for animation.

Make sure that SkGifDecoder::Decode can still succeed with a nullptr
for DecodeContext.

Bug: 311332788
Change-Id: I7a9e0fc79fa7fb246614a83fb3a6fbd0f30c0c8b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/790318
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>
Reviewed-by: John Reck <jreck@google.com>