[graphite] Remove conditional uploads for the text atlas.
This is step one in fixing the below bug. Part of the solution is not
updating the TokenTracker when a child layer flushes alone. This gives
slightly less efficient usage but guarantees that the parent will
not overwrite Plots that it has draws for but that haven't been
flushed yet.

However, doing this exposes an issue with conditional uploads, where a
follow-on upload was not being added to the CommandBuffer because its
token value was the same as the previous one. We aren't going to use
this system to solve out-of-order Recording playback, so we might as
well switch back to the "Iron Fist" method so this bug can be addressed.

Bug: b/316372536
Change-Id: I98d47739afeccd34d1b0913569677b68791cd89d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/790076
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>