Reland "Cache the one SkEmptyTypeface"
This reverts commit 10e7ee119ea9731e0924b4f80cea0fc24ed32182.

All instances of SkEmptyTypeface are equivelent in that they have no
glyphs and no state. As a result, have its factory function always
return a reference to a single cached instance.

For now this single cached instance must be created on first access. In
the future with some work it may be possible for this typeface to be
constexpr and created at compile time.

Prevent any issues static destruction order by allowing this instance to
"leak" with SkNoDestructor.

Bug: chromium:1511268
Change-Id: Iccf598f4eb2b8e9d84a08f923390f2cd429e98f4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/791341
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>