Update RTFlip handling to avoid caps-bit checks in the front-end.
Previously, we had a boolean in the Program::Interface which meant
"this program needs RTFlip," and the caps bit for CanUseFragCoord
influenced the final state of the boolean.

Now, we have a bitfield in the Program::Interface which enumerates
the things which can influence the existence of RTFlip (frag
coords, clockwise, and derivatives). In the calling code, if any
bits are set, we know we need RTFlip; however, we can ignore
the RTFlag_FragCoord bit if CanUseFragCoords is disabled. Note that
CanUseFragCoords is only disabled in GL, so this logic only exists
there.

Bug: b/316916102
Change-Id: Ic721d31682e6b3ff44fdcdfdaf9fc29b3dea1da3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/792036
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>