Move Compiler::convertIdentifier method into SymbolTable.
This method originally lived in the IRGenerator and was migrated
into the compiler out of convenience once the IRGenerator went
away (see http://review.skia.org/459936,
https://skia-review.googlesource.com/c/skia/+/459936/comment/6a16e868_5bb35ed7/).

Upon reflection, this belongs in SymbolTable. (An alternative
factoring: it could live in Symbol, and the lookup can be done
manually by the caller, but that's a more complex change.)

Removing this will allow us to decouple the SPIR-V code generator
(which instantiates symbols) from the SkSLCompiler entirely.

Change-Id: I82743cbb0a33a4cfcef9eb0782abe7255874389d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/792837
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>