Disable lax-vector-conversions in clang
A couple years ago, clang tried to switch the default behavior of this
switch, and Skia's code was the only failing instance in Chrome. It's
expected that clang may switch the default again in the future. In any
case, these implicit conversions are questionable, so it's better that
we make them explicit.

Also, GCC is strict about this behavior by default, so fixing this will
be necessary to get RP support in Firefox.

Bug: chromium:1042470
Bug: b/316617002
Change-Id: I08abec38b1d5ef28dbad076cff42823ac4725204
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/793316
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>