Update bots' NDK to r26b
Our NDK was several years out of date, so it was missing newer APIs.
The last time we tried an update, it caused ASAN issues on the tree:
    https://skia-review.googlesource.com/c/skia/+/439336

Like that attempt, this changes the Mac create script, because the
NDK is now distributed as a DMG.

Critically, this moves our Android bots from clang 9 (!) to 17.

Change-Id: I2bc4204e63dc1d3b41b4f04bd1d2e62713fdedc9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/793577
Reviewed-by: Ben Wagner <bungeman@google.com>